build: add 'gradle dev' task to run in headless mode without building javascript; moved cleandb and generateTypes to modules to make them start faster again bypassing javascript build'
