+++ backend/molgenis-emx2-run/src/test/java/org/molgenis/emx2/AToolToGenerateTypeScriptTypesTest.java
@@ -6,6 +6,7 @@ import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.nio.charset.Charset;
 import org.junit.jupiter.api.Test;
+import org.molgenis.emx2.typescript.AToolToGenerateTypeScriptTypes;
 
 class AToolToGenerateTypeScriptTypesTest {
