+++ backend/molgenis-emx2-sql/build.gradle
@@ -6,3 +6,10 @@ dependencies {
     testImplementation project(':backend:molgenis-emx2-datamodels')
     testImplementation project(':backend:molgenis-emx2-tasks')
 }
+
+//tool should be here otherwise requires build of unnecessary modules
+task cleandb(type: JavaExec) {
+    group = 'application'
+    classpath = sourceSets.main.runtimeClasspath
+    main = 'org.molgenis.emx2.sql.AToolToCleanDatabase'
+}
