+++ backend/molgenis-emx2-typescript/build.gradle
@@ -4,3 +4,9 @@ dependencies {
     implementation project(':backend:molgenis-emx2-datamodels')
     testImplementation project(':backend:molgenis-emx2-tasks')
 }
+
+task generateTypes(type: JavaExec) {
+    group = 'application'
+    classpath = sourceSets.main.runtimeClasspath
+    main = 'org.molgenis.emx2.typescript.AToolToGenerateTypeScriptTypes'
+}
