+++ docs/molgenis/run_updates.md
@@ -16,7 +16,7 @@ below. Caution: this will delete all MOLGENIS generated roles (MG_ROLE*, MG_USER
 
 Download the appropriate *.jar release and run command:
 
-```java -cp molgenis-emx2-<version>-all.jar org.molgenis.emx2.AToolToCleanDatabase```
+```java -cp molgenis-emx2-<version>-all.jar org.molgenis.emx2.sql.AToolToCleanDatabase```
 
 If you have the sourcecode you can also run:
 ```gradle cleandb```
